created order api

+++ app/api/product/[slug]/route.ts
@@ -1,4 +1,4 @@
-const db=`${process.env.DB_URL}/${process.env.DB_NAME}`
+const db = `${process.env.DB_URL}/${process.env.DB_NAME}`;
 import mongoose from "mongoose";
 mongoose.connect(db);
 
@@ -6,11 +6,13 @@ import serverCatchError from "@/lib/server-catch-error";
 import { NextRequest, NextResponse as res } from "next/server";
 import ProductModel from "@/models/product.model";
 import SlugInterface from "@/interface/slug.interface";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../../auth/[...nextauth]/route";
 
 export const GET = async (req: NextRequest, context: SlugInterface) => {
   try {
     const { slug } = await context.params;
-    console.log(slug)
+    console.log(slug);
     const product = await ProductModel.findOne({ slug });
     if (!product)
       return res.json(
@@ -25,6 +27,12 @@ export const GET = async (req: NextRequest, context: SlugInterface) => {
 
 export const PUT = async (req: NextRequest, { params }: SlugInterface) => {
   try {
+    const session = await getServerSession(authOptions);
+    if (!session) return res.json({ message: "Unauthorized" }, { status: 401 });
+
+    if (session.user.role !== "admin")
+      return res.json({ message: "Unauthorized" }, { status: 401 });
+
     const { slug: id } = await params;
     const body = await req.json();
     const product = await ProductModel.findByIdAndUpdate(id, body, {
@@ -43,6 +51,11 @@ export const PUT = async (req: NextRequest, { params }: SlugInterface) => {
 
 export const DELETE = async (req: NextRequest, { params }: SlugInterface) => {
   try {
+    const session = await getServerSession(authOptions);
+    if (!session) return res.json({ message: "Unauthorized" }, { status: 401 });
+
+    if (session.user.role !== "admin")
+      return res.json({ message: "Unauthorized" }, { status: 401 });
     const { slug: id } = await params;
 
     const product = await ProductModel.findByIdAndDelete(id);
